requirements.txt
tqdm
matplotlib
numpy == 1.19.5
scipy == 1.1.0
Pillow == 5.2.0
pandas == 0.24.2
shapely == 1.7.1
geopandas == 0.8.2
geojson == 2.5.0
scikit-learn == 0.22
statsmodels == 0.9.0
plotly == 4.14.3
pyshp == 2.1.2
torch == 1.1.0
torchvision == 0.2.2